requirements.txt
Flask==2.3.2
Flask-CORS==4.0.1
supabase==2.0.0
langchain==0.3.0
langchain-google-genai==2.0.0
google-generativeai>=0.7.0,<0.8.0
pdfminer.six==20250506
python-dotenv==1.1.1
psycopg2-binary==2.9.9
werkzeug==2.3.7